Changed config definition for config rows
